issue:#13:refactor
